Optimize permission check
If the namespace is not enabled for "secure page assignments", we don't
need to check if the title exists (actually we don't need to do this at
all, as `PageAssignmentsFactory::newFromTargetTitle` does it
implicitly). By putting the check for existence _after_ the check for
enabled namespace, we can avoid unnecessary database queries by
`Title::exist`.

This improved performance for SMW queries with ~5000 results by factor
2.

ERM38583

[5.1.x]

Change-Id: I1427b8bc693d351ffcce85041e2253ef55f6f938